Keep this branch's selective-CI model (sweeps and conformance windows run
as selectable shards, so a PR runs only what its changes touch and a merge
does not re-run what the PR already validated) and make master's additions
consistent with it:

- sonarcloud.yml: keep the deletion of the standalone
  parameter-enumeration-sweep and model-shape-conformance-windows jobs,
  and drop the six now-orphaned `needs:` edges master added to
  test-regression-analysis, the validation gate and ci-test-analysis. A
  `needs:` on an undefined job is an invalid workflow. Master's
  select-shards outputs (requires_sweeps / requires_shapes /
  parameter_matrix) are kept and still drive shard selection.
- Assert-CiGate.ps1: keep this branch's requirement set. The -RequiresSweeps
  / -RequiresShapes parameters master added are retained and still accepted
  (defaulting to 'true'); only the clause requiring the two deleted jobs is
  dropped.
- Test-CiImpactWorkflowReview.ps1: keep this branch's manifest-based
  negative controls and its ManifestContent handling; master's job-gate
  controls bound to the two deleted jobs. Take master's budget-shrunk
  mutation (timeout-minutes 5), a clear shrink from the actual 50.
- Test-TestImpactEndToEnd.ps1: take master's comment, which matches the
  assertions below it (rerun Always,Beta; import Alpha).

Verified: no reference to either deleted job remains, every `needs:`
target resolves to a defined job, and all three PowerShell files parse.

+++ src/Preprocessing/FeatureSelection/Filter/Relief/MultiSURF.cs
@@ -136,7 +136,7 @@ public class MultiSURF<T> : TransformerBase<T, Matrix<T>, Matrix<T>>
                 {
                     double hit_val = NumOps.ToDouble(data[hitIdx, j]);
                     double diffHit = Math.Abs(ri_val - hit_val) / range;
-                    _featureWeights[j] -= diffHit / (n * Math.Max(1, nearHits.Count));
+                    _featureWeights[j] -= diffHit / ((double)n * Math.Max(1, nearHits.Count));
                 }
 
                 // Contribution from misses
@@ -144,7 +144,7 @@ public class MultiSURF<T> : TransformerBase<T, Matrix<T>, Matrix<T>>
                 {
                     double miss_val = NumOps.ToDouble(data[missIdx, j]);
                     double diffMiss = Math.Abs(ri_val - miss_val) / range;
-                    _featureWeights[j] += diffMiss / (n * Math.Max(1, nearMisses.Count));
+                    _featureWeights[j] += diffMiss / ((double)n * Math.Max(1, nearMisses.Count));
                 }
             }
         }
